requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
requests==2.32.3
APScheduler==3.10.4
reportlab==4.2.2
arabic-reshaper==3.0.0
python-bidi==0.4.2
gunicorn==23.0.0
psycopg2-binary==2.9.9
google-api-python-client==2.134.0
google-auth==2.32.0
